fsimx8mm/mn/mp: Add support for U-ATF and U-TEE
On fsimx8mm/mn/mp, ATF and TEE have been part of NBoot in the past. This
had the disadvantage that ATF and TEE could not be updated without updating
the rest of NBoot, too. Which meant that customers were forced to handle
an own NBoot if they wanted to have an own ATF or TEE version.

This commit adds the possibility to optionally group ATF/TEE with U-Boot.
By setting CONFIG_FS_ADD_ATF_TO_UBOOT, a new target uboot-atf.fs is built
that also holds ATF (and TEE if activated). Of course these files have to
be present in board/F+S/NXP-Firmware in this case.

This new image has a new F&S header U-BOOT-ATF and groups the U-ATF,
the optional U-TEE and the unmodified U-BOOT sub-images. U-ATF and U-TEE
stand for "User ATF" and "User TEE" in contrast to the regular now-called
System ATF and System TEE images that are still part of every NBoot image.
You can also think of U-ATF/U-TEE as images being part of U-Boot.

   +------------------------------+
   | U-BOOT-ATF (arch)            |
   |   +--------------------------+
   |   | U-ATF (arch)             |
   |   +--------------------------+
   |   | U-TEE (arch) (optional)  |
   |   +--------------------------+
   |   | U-BOOT (arch)            |
   +------------------------------+

As long as only regular U-BOOT images are installed, the system behaves as
before. But at the moment where there is a U-BOOT-ATF image installed, the
U-ATF and U-TEE images replace the previous system images. From then on,
the regular ATF and TEE images are ignored when storing a new NBoot version,
keeping the user versions. Which means U-ATF and U-TEE now logically belong
to the user and can be updated together with U-Boot by saving U-BOOT-ATF
images. However it is still possible to just update U-Boot with a regular
U-BOOT image.

By using option -s (for "System-ATF") in command fsimage save, this behaviour
can be reverted. If -s is given when saving NBoot, then the ATF and TEE of
NBoot are saved again, restoring the original behavior. If option -s is given
when saving an U-BOOT-ATF image, the U-ATF and U-TEE given there are
ignored, which means such an image can also be used to just update U-Boot,
keeping the System ATF/TEE images.

Remark:

When downloading software with uuu, the board gets NBoot and U-Boot in one
go, which means the behavior stays as before. It uses the System ATF/TEE
from NBoot and cannot accept a U-BOOT-ATF image.

The following changes are needed to achieve this:

- Add option CONFIG_FS_ADD_ATF_TO_UBOOT to board/F+S/common/Kconfig.
- Add rules to build U-BOOT-ATF to the global Makefile.
- Add handling for U-ATF and U-TEE in fs_image_load_system_copy() in
  board/F+S/common/fs_image_spl.c.

The following changes are in cmd/fsimage.c:

- Add function fs_image_is_u_atf() to verify if the stored ATF is a
  System ATF or a User U-ATF. TEE is always handled exactly the same
  as ATF, so no extra check is needed there.
- Add Option -s to do_fsimage_save(). Add explanation to fsimage
  command help text.
- In fsimage_imx8_save(), add handling for U-ATF/U-TEE, once when saving
  NBoot, but also when loading U-Boot in case it has to be relocated.
- In do_fsimage_save_uboot(), handle U-ATF/U-TEE if U-BOOT-ATF image is
  found to be stored.
- In fsimage_imx8_load(), skip loading ATF/TEE if U-ATF/U-TEE is present in
  flash memory.
- In fs_image_load_uboot(), build a U-BOOT-ATF instead of U-BOOT image in
  RAM if U-ATF/U-TEE are present in flash memory.