fsimage.c: Clear secondary flag of SPL sub-image in fsimage load
When loading NBoot with fsimage load, SPL showed a difference in the two
copies of the image. Inspection showed that this is the secondary flag that
is set in the second copy. Clear the flag in fs_image_load_image() when
loading the SPL image before comparing the two copies.

Simplify the code for fs_image_set_spl_secondary_bit().