Fix warning from a2f4736d9407b350b0f973865b376c5583a2e6eb
The commit a2f4736d9407b350b0f973865b376c5583a2e6eb introduced functionality that
added support for signed images when using the fsimage list command. For this, the
function "fs_image_check_for_ivt" was added that takes an "uintptr_t *" as its second
argument. The Code that calls it gives an "uint32_t *", causing a warning.  The function
is only called in this one place. Change the second Argument of the function to better fit
the overlaying structure.